+++ HISTORY.md
@@ -1,3 +1,7 @@
+### v0.1.0 Imports insertion and rxjs/most support 
+    * imports of `Subject` object are added if not found
+    * added initial most.js and rxjs5 support    
+
 ### v0.0.1 Initial pre-release 
     * only rx@4.x.x 
     * requires `Rx` or `Subject` reference